+++ src/components/sbb-button/sbb-button.stories.js
@@ -1,6 +1,5 @@
 import images from '../../global/images';
 import { SbbColorWhiteDefault } from '@sbb-esta/lyne-design-tokens';
-import events from './sbb-button.events.ts';
 import { h } from 'jsx-dom';
 import readme from './readme.md';
 
@@ -93,7 +92,7 @@ const isStatic = {
   control: { type: 'boolean' },
 };
 
-const idValue = {
+const buttonId = {
   control: {
     type: 'text',
   },
@@ -232,7 +231,7 @@ const defaultArgTypes = {
   negative,
   size,
   static: isStatic,
-  'id-value': idValue,
+  'button-id': buttonId,
   'icon-name': iconName,
   href,
   target,
@@ -256,7 +255,7 @@ const defaultArgs = {
   negative: false,
   size: size.options[0],
   static: false,
-  'id-value': undefined,
+  'button-id': undefined,
   'icon-name': 'arrow-right-small',
   href: undefined,
   target: undefined,
@@ -534,7 +533,7 @@ export default {
   ],
   parameters: {
     actions: {
-      handles: [events.click],
+      handles: ['click'],
     },
     backgrounds: {
       disable: true,