+++ src/components/sbb-link/sbb-link.stories.js
@@ -10,7 +10,6 @@ import {
 } from '@sbb-esta/lyne-design-tokens';
 import { h } from 'jsx-dom';
 import readme from './readme.md';
-import events from './sbb-link.events';
 
 const wrapperStyle = (context) => {
   if (!context.args.negative) {
@@ -86,7 +85,7 @@ const isStatic = {
   control: { type: 'boolean' },
 };
 
-const idValue = {
+const linkId = {
   control: {
     type: 'text',
   },
@@ -235,7 +234,7 @@ const defaultArgTypes = {
   negative,
   'text-size': textSize,
   static: isStatic,
-  'id-value': idValue,
+  'link-id': linkId,
   'icon-name': iconName,
   'icon-placement': iconPlacement,
   href,
@@ -260,7 +259,7 @@ const defaultArgs = {
   negative: false,
   'text-size': textSize.options[1],
   static: false,
-  'id-value': undefined,
+  'link-id': undefined,
   'icon-name': undefined,
   'icon-placement': iconPlacement.options[0],
   href: 'https://github.com/lyne-design-system/lyne-components',
@@ -499,7 +498,7 @@ export default {
   ],
   parameters: {
     actions: {
-      handles: [events.click],
+      handles: ['click'],
     },
     docs: {
       extractComponentDescription: () => readme,