+++ src/components/sbb-link/sbb-link.tsx
@@ -1,18 +1,7 @@
-import {
-  Component,
-  ComponentInterface,
-  Element,
-  Event,
-  EventEmitter,
-  h,
-  JSX,
-  Listen,
-  Prop,
-  State,
-} from '@stencil/core';
+import { Component, ComponentInterface, Element, h, JSX, Listen, Prop, State } from '@stencil/core';
 import {
   ButtonType,
-  forwardHostClick,
+  forwardHostEvent,
   LinkButtonProperties,
   LinkButtonRenderVariables,
   LinkTargetType,
@@ -58,7 +47,7 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
   @Prop({ attribute: 'static', mutable: true, reflect: true }) public isStatic = false;
 
   /** Pass in an id, if you need to identify the inner element. */
-  @Prop() public idValue?: string;
+  @Prop() public linkId?: string;
 
   /**
    * The icon name we want to use, choose from the small icon variants
@@ -111,14 +100,6 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
    */
   @Prop() public accessibilityHaspopup: PopupType | undefined;
 
-  /** Emits the event on button click. */
-  @Event({
-    bubbles: true,
-    composed: true,
-    eventName: 'sbb-link-button_click',
-  })
-  public click: EventEmitter;
-
   /** This will be forwarded as aria-label to the relevant nested element. */
   @Prop() public accessibilityLabel: string | undefined;
 
@@ -141,6 +122,13 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
     // Check if the current element is nested in a form.
     this._closestForm = hostContext('form', this._element) as HTMLFormElement;
     this._namedSlots = queryAndObserveNamedSlotState(this._element, this._namedSlots);
+
+    // Forward focus call to action element
+    this._element.focus = (options: FocusOptions) => this._actionElement().focus(options);
+  }
+
+  private _actionElement(): HTMLElement {
+    return this._element.shadowRoot.firstElementChild as HTMLElement;
   }
 
   /**
@@ -150,8 +138,6 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
     if (this.disabled || this.isStatic) {
       return;
     }
-
-    this.click.emit();
 
     if (!this._closestForm || this.type !== 'submit') {
       return;
@@ -169,12 +155,8 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
     if (this.disabled) {
       event.preventDefault();
       event.stopImmediatePropagation();
-    } else {
-      forwardHostClick(
-        event,
-        this._element,
-        this._element.shadowRoot.firstElementChild as HTMLElement // button or a element
-      );
+    } else if (!this.isStatic) {
+      forwardHostEvent(event, this._element, this._actionElement());
     }
   }
 
@@ -193,7 +175,7 @@ export class SbbLink implements LinkButtonProperties, ComponentInterface {
     // See https://github.com/ionic-team/stencil/issues/2703#issuecomment-1050943715 on why form attribute is set with `setAttribute`
     return (
       <TAG_NAME
-        id={this.idValue}
+        id={this.linkId}
         class="sbb-link"
         {...attributes}
         ref={(btn) => this.form && btn?.setAttribute('form', this.form)}