+++ src/components/sbb-timetable-row/sbb-timetable-row.spec.ts
@@ -2,13 +2,15 @@ import { newSpecPage } from '@stencil/core/testing';
 import { SbbTimetableRow } from './sbb-timetable-row';
 import { config } from './sbb-timetable-row.sample-data';
 
+const now = new Date('2022-08-16T15:00:00Z').valueOf();
+
 describe('sbb-timetable-row', () => {
   describe('sbb-timetable-row with config', () => {
     it('renders component with config', async () => {
       const page = await newSpecPage({
         components: [SbbTimetableRow],
         html: `
-            <sbb-timetable-row>
+            <sbb-timetable-row data-now="${now}">
             </sbb-timetable-row>
         `,
       });
@@ -16,7 +18,7 @@ describe('sbb-timetable-row', () => {
       page.rootInstance.price = config.price;
       await page.waitForChanges();
       expect(page.root).toEqualHtml(`
-      <sbb-timetable-row>
+      <sbb-timetable-row data-now="1660662000000">
       <mock:shadow-root>
         <sbb-card>
           <sbb-card-badge appearance="primary" isdiscount="" price="12" slot="badge" text="ab CHF"></sbb-card-badge>
@@ -30,7 +32,7 @@ describe('sbb-timetable-row', () => {
                 Direction Luzern
               </p>
             </div>
-            <sbb-pearl-chain-time arrivaltime="2022-10-28T21:16:00+02:00" arrivalwalk="5" departuretime="2022-10-28T02:48:00+02:00" departurewalk="8"></sbb-pearl-chain-time>
+            <sbb-pearl-chain-time arrivaltime="2022-10-28T21:16:00+02:00" arrivalwalk="5" departuretime="2022-10-28T02:48:00+02:00" departurewalk="8" data-now="1660662000000"></sbb-pearl-chain-time>
             <div class="sbb-timetable__row-footer">
               <span class="sbb-timetable__row-quay--changed">
                 <span class="sbb-screenreaderonly">
@@ -96,11 +98,11 @@ describe('sbb-timetable-row', () => {
     it('renders loading state', async () => {
       const page = await newSpecPage({
         components: [SbbTimetableRow],
-        html: `<sbb-timetable-row loading-trip="true" />`,
+        html: `<sbb-timetable-row loading-trip="true" data-now="${now}"/>`,
       });
       page.rootInstance.config = { 'loading-trip': true };
       expect(page.root).toEqualHtml(`
-        <sbb-timetable-row loading-trip="true">
+        <sbb-timetable-row loading-trip="true" data-now="1660662000000">
           <mock:shadow-root>
             <button class="sbb-loading" disabled="">
               <div class="sbb-loading__wrapper">